By default, make cards fill available width now (#55)
The desired UI on Mac is:
- For cards that are in a dialog (e.g. project and app
preferences - most of them) make the card fill available
width (with a minimum). If the dialog is resized, the
card grows horizontally.
- For cards that are in a doc windows (e.g. Android
manifest and iOS PList), give them a fixed width since
the doc window can be very wide (esp. with a wide
screen monitor) and filling it all looks funny.